requirements.txt
amqp==5.1.1
aniso8601==9.0.1
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.5.2
asttokens==2.0.5
async-timeout==4.0.2
attrs==21.4.0
backcall==0.2.0
beautifulsoup4==4.11.1
billiard==3.6.4.0
bleach==5.0.0
celery==5.2.6
certifi==2022.5.18.1
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
cycler==0.11.0
debugpy==1.6.0
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.13
Django==4.0.4
djangorestframework==3.13.1
entrypoints==0.4
executing==0.8.3
fastjsonschema==2.15.3
Flask==2.1.2
Flask-RESTful==0.3.9
fonttools==4.33.3
greenlet==1.1.2
idna==3.3
importlib-metadata==4.11.4
install==1.3.5
ipykernel==6.13.0
ipython==8.3.0
ipython-genutils==0.2.0
ipywidgets==7.7.0
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
jsonschema==4.5.1
jupyter==1.0.0
jupyter-client==7.3.1
jupyter-console==6.4.3
jupyter-core==4.10.0
jupyterlab-pygments==0.2.2
jupyterlab-widgets==1.1.0
kiwisolver==1.4.2
kombu==5.2.4
MarkupSafe==2.1.1
matplotlib==3.5.2
matplotlib-inline==0.1.3
mistune==0.8.4
nbclient==0.6.3
nbconvert==6.5.0
nbformat==5.4.0
nest-asyncio==1.5.5
notebook==6.4.11
numpy==1.22.4
packaging==21.3
pandas==1.4.2
pandocfilters==1.5.0
parso==0.8.3
pickleshare==0.7.5
Pillow==9.1.1
prometheus-client==0.14.1
prompt-toolkit==3.0.29
psutil==5.9.1
psycopg2-binary==2.9.3
pure-eval==0.2.2
pycparser==2.21
Pygments==2.12.0
pyparsing==3.0.9
pyrsistent==0.18.1
python-dateutil==2.8.2
pytz==2022.1
pywin32==304
pywinpty==2.0.5
pyzmq==23.0.0
qtconsole==5.3.0
QtPy==2.1.0
redis==4.3.1
requests==2.27.1
schedule==1.1.0
scipy==1.8.1
Send2Trash==1.8.0
six==1.16.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.36
sqlparse==0.4.2
stack-data==0.2.0
terminado==0.15.0
tinycss2==1.1.1
tornado==6.1
traitlets==5.2.1.post0
tzdata==2022.1
urllib3==1.26.9
vine==5.0.0
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.1.2
widgetsnbextension==3.6.0
wrapt==1.14.1
zipp==3.8.0
